"LA.UM.9.14.r1-20700-LAHAINA.QSSI13.0"

# By Krishna Kishor Jha (2) and others
# Via Gerrit - the friendly Code Review server (5) and Linux Build Service Account (1)
* tag 'LA.UM.9.14.r1-20700-LAHAINA.QSSI13.0':
  audio-hal: add missing modules in audio_effects.xml
  hal: changing mtmx coeffs variable type to uint32_t
  audio-hal: automotive LL capture usecase changes blocking with flag
  config: added audio effects config for elite architecture
  config: added audio effect libraries and helpers.
  audio-hal: Fix build error with VA and AOSP tinycompress

+++ hal/audio_extn/audio_extn.c
@@ -635,7 +635,7 @@ static int set_custom_mtmx_output_channel_map(struct audio_device *adev,
     struct mixer_ctl *ctl = NULL;
     char mixer_ctl_name[128] = {0};
     int ret = 0;
-    long channel_map[AUDIO_MAX_DSP_CHANNELS] = {0};
+    int channel_map[AUDIO_MAX_DSP_CHANNELS] = {0};
 
     ALOGV("%s channel_count %d", __func__, ch_count);
